+++ lib/Db/Vote.php
@@ -44,7 +44,7 @@ use OCP\AppFramework\Db\Entity;
  * @method string getVoteAnswer()
  * @method void setVoteAnswer(string $value)
  */
-class Vote extends Entity implements JsonSerializable {
+class Vote extends EntityWithUser implements JsonSerializable {
 	public const TABLE = 'polls_votes';
 
 	/** @var int $pollId */
@@ -65,18 +65,18 @@ class Vote extends Entity implements JsonSerializable {
 	/** @var string $voteAnswer */
 	protected $voteAnswer;
 
-	/** @var IUserManager */
-	private $userManager;
+	// /** @var IUserManager */
+	// private $userManager;
 
-	/** @var ShareMapper */
-	private $shareMapper;
+	// /** @var ShareMapper */
+	// private $shareMapper;
 
 	public function __construct() {
 		$this->addType('id', 'int');
 		$this->addType('pollId', 'int');
 		$this->addType('voteOptionId', 'int');
-		$this->userManager = Container::queryClass(IUserManager::class);
-		$this->shareMapper = Container::queryClass(ShareMapper::class);
+		// $this->userManager = Container::queryClass(IUserManager::class);
+		// $this->shareMapper = Container::queryClass(ShareMapper::class);
 	}
 
 	/**
@@ -88,48 +88,51 @@ class Vote extends Entity implements JsonSerializable {
 			'pollId' => $this->getPollId(),
 			'optionText' => $this->getVoteOptionText(),
 			'answer' => $this->getVoteAnswer(),
-			'user' => [
-				'userId' => $this->getPublicUserId(),
-				'displayName' => $this->getDisplayName(),
-				'isNoUser' => $this->getIsNoUser()
-			],
+			'user' => $this->getUser(),
 		];
 	}
 
-	public function getDisplayName(): string {
-		if ($this->getIsNoUser()) {
-			// get displayName from share
-			try {
-				$share = $this->shareMapper->findByPollAndUser($this->getPollId(), $this->userId);
-			} catch (ShareNotFoundException $e) {
-				// use fake share
-				$share = $e->getReplacement();
-			}
-			return $share->getDisplayName();
-		}
+	// public function getIsNoUser(): bool {
+	// 	return !($this->userManager->get($this->userId) instanceof IUser);
+	// }
 
-		return $this->userManager->get($this->userId)->getDisplayName();
-	}
+	// public function getDisplayName(): string {
+	// 	if ($this->getIsNoUser()) {
+	// 		// get displayName from share
+	// 		try {
+	// 			$share = $this->shareMapper->findByPollAndUser($this->getPollId(), $this->userId);
+	// 		} catch (ShareNotFoundException $e) {
+	// 			// use fake share
+	// 			$share = $e->getReplacement();
+	// 		}
+	// 		return $share->getDisplayName();
+	// 	}
 
-	public function getIsNoUser(): bool {
-		return !($this->userManager->get($this->userId) instanceof IUser);
-	}
+	// 	return $this->userManager->get($this->userId)->getDisplayName();
+	// }
 
-	private function getPublicUserId() {
-		if (!$this->getUserId()) {
-			return '';
-		}
+	// private function getPublicUserId() {
+	// 	if (!$this->getUserId()) {
+	// 		return '';
+	// 	}
 
-		if ($this->publicUserId) {
-			return $this->publicUserId;
-		}
+	// 	if ($this->publicUserId) {
+	// 		return $this->publicUserId;
+	// 	}
 
-		return $this->getUserId();
-	}
+	// 	return $this->getUserId();
+	// }
 
-	public function generateHashedUserId() {
-		$this->publicUserId = hash('md5', $this->getUserId());
-	}
+	// public function generateHashedUserId() {
+	// 	$this->publicUserId = hash('md5', $this->getUserId());
+	// }
 
+	// public function getUser(): array {
+	// 	return [
+	// 		'userId' => $this->getPublicUserId(),
+	// 		'displayName' => $this->getDisplayName(),
+	// 		'isNoUser' => $this->getIsNoUser(),
+	// 	];
+	// }
 }
